// src/lib/api.types.ts
export interface components {
    schemas: {
// ...
        FormQuestionInput: {
            /** Id */
            id: string;
            /** Text */
            text: string;
            /**
             * Answer Type
             * @enum {string}
             */
            answer_type: "text" | "select";
            /** Answer Select Options */
            answer_select_options?: string[] | null;
            /** Answer Select Multiple */
            answer_select_multiple?: boolean | null;
            /** Required */
            required: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}